fix paths that contain (but not are) symlinks
if part of the path contains a symbolic link, the readlink command will
identify it as a regular file. it only tells you if the "last" part of
the path is a symbolic link

fix this by resolving all paths with realpath() and checking if
different. the full symbolic link is stored now instead of just suffix