Reorganize dependency tracking.
Dependencies are now organized using maps and sets. This ensures that no
duplicate entries occur. As there are more lookups than direct acesses and
real iterations this should be running faster than the original approach.
Internally it reduces the worst case complexity from n^4 to n^2*log(n)^2
where n is the number of files. The worst case complexity of external shell
calls is reduced from n^3 to n^2.

Unhandled files are saved in a queue so that we can merge Dependencies directly.

This is not expected to solve all issues with the program so far.