Add -xr or --fix-recursiver argument to fix all objects in folder
Sometimes, when bundling apps with lots of plug-ins, it is cumbersome
to list all these plug-ins or modules directly when passing them to
the `dylibbundler` tool.  With this new argument, the tool can find by
itself all files to fix in a given tree.

For now, it does so by trying to detect whether it's a Mach-O object
file.  Alternatively we could use a different check (e.g. looking for
certain extensions, like `.so` files).